Fix default agent version
The value for the agent is never null. If it is not run
with maven then the value will be ${project.version}. A check is
added. If that is the case then the value is replaced with the
default value. In addition the IT test is changed to expect the
default value.